fix: Align all entity mappings, DTOs, services, and controllers with actual database schema
Major changes:
- User: PK changed to AccountLogIn, password field to AccountPassWord, permissions to nvarchar(600) strings, IsActive to Stop, added AddedBy/ModifiedBy/ModifyDate/Signature/Dept
- Branch: Column names corrected (BranchNameEn, Telephone, EMail), AccountingCode to nvarchar(2), added BranchCode/Agent/StartAt/Logo/MainCenter/CompanyOffice/SystemURI
- Customer: CustNo changed to bigint, NationalId to IDNo, PassportNo to DrCardNo, removed DateOfBirth/Branch, added Key/CustTP/CustNat/RecDate/FaxNo/Bank/BankAcc
- Agent: Replaced non-existent AgentFile with AgentsCommisions table (agents are branches with Agent=1 in BranchInfo)
- Claim: Composite PK (ClmNo, SubIns, Branch), decimal fields to double
- ClaimEstimation: Composite PK (TPID, ClmNo, Date)
- ClaimSettlement: Composite PK (ClmNo, TPID, No), added DAILYNUM/DAILYDTE/SerNo
- Receipt: Maps to ACCMOVE table (not ReceiptFile), PK is SerNo
- LogEntry: Maps to LogData table (not LogFile)
- Journal: Corrected MainJournal and JournalDetail mappings with composite PKs
- Account: Removed Branch field, preserved TranscationAcc typo from DB
- Removed non-existent tables (Notifications, Treaties, SubSystem)

Build: 0 errors, 0 warnings

+++ InsuranceAPI/src/InsuranceAPI.Application/Services/Interfaces/ICustomerService.cs
@@ -5,9 +5,9 @@ namespace InsuranceAPI.Application.Services.Interfaces;
 
 public interface ICustomerService
 {
-    Task<ApiResult<CustomerDto>> GetByIdAsync(int custNo);
+    Task<ApiResult<CustomerDto>> GetByIdAsync(long custNo);
     Task<ApiResult<PaginatedResult<CustomerDto>>> GetAllAsync(int page = 1, int pageSize = 20, string? search = null);
     Task<ApiResult<CustomerDto>> CreateAsync(CreateCustomerRequest request);
-    Task<ApiResult<CustomerDto>> UpdateAsync(int custNo, UpdateCustomerRequest request);
-    Task<ApiResult<bool>> DeleteAsync(int custNo);
+    Task<ApiResult<CustomerDto>> UpdateAsync(long custNo, UpdateCustomerRequest request);
+    Task<ApiResult<bool>> DeleteAsync(long custNo);
 }
